+++ dotNetInstallerLib/InstalledCheckDirectory.h
@@ -1,4 +1,6 @@
 #pragma once
+
+#include "XmlAttribute.h"
 #include "InstalledCheck.h"
 
 class InstalledCheckDirectory : public InstalledCheck
@@ -6,6 +8,10 @@ class InstalledCheckDirectory : public InstalledCheck
 public:
 	// full path to the directory to check
 	std::wstring path;
+	// The type of comparison.
+	XmlAttribute comparison;
+	// The version number to compare against.
+	XmlAttribute version;
 public:
     InstalledCheckDirectory();
     void Load(tinyxml2::XMLElement * node);