+++ dotNetInstallerToolsLib/FileUtil.h
@@ -59,18 +59,4 @@ namespace DVLib
 	// load resource data from a module
 	template<class T>
 	std::vector<T> LoadResourceData(HMODULE h, const std::wstring& resource, const std::wstring& type);
-
-	// 4-part file version
-	struct FileVersion
-	{
-		int major;
-		int minor;
-		int build;
-		int rev;
-	};
-
-	FileVersion wstring2fileversion(std::wstring version);
-	std::wstring fileversion2wstring(const FileVersion& version);
-	// compare a Windows-specific 4-part version
-	int CompareVersion(const std::wstring& l, const std::wstring& r);
 }