+++ dotNetInstallerToolsLib/Logger.h
@@ -1,0 +1,12 @@
+#pragma once
+#include <string>
+
+namespace DVLib
+{
+    class Logger
+    {
+    public:
+        virtual ~Logger() {}
+        virtual void Log(const std::wstring& message) = 0;
+    };
+}